refactor: split 3 large files to ≤150 LOC via team-agents (3 Sonnet parallel)
Three files over 200 LOC split by a 3-agent team working in parallel:

bud/impl.ts (382 → 4 files):
  impl.ts (143) — cmdBud orchestrator + BudOpts interface
  bud-repo.ts (41) — GitHub repo creation + ghq clone
  bud-init.ts (128) — vault init, CLAUDE.md, fleet config, birth note
  bud-wake.ts (128) — soul-sync, git commit, sync_peers, wake, split

soul-sync/impl.ts (403 → 3 files):
  impl.ts (163) — cmdSoulSync + cmdSoulSyncProject
  resolve.ts (83) — resolveOraclePath, resolveProjectSlug, findOracleForProject
  sync-helpers.ts (150) — syncDir, findPeers, syncOracleVaults, syncProjectVault

config.ts (542 → 6 files):
  config.ts (6) — barrel re-exports only
  config-types.ts (102) — interfaces + D constants
  config-validate.ts (142) — validateBasicFields + validateConfigShape
  config-validate-ext.ts (138) — validateExtFields
  config-load.ts (110) — detectGhqRoot, DEFAULTS, cache, loadConfig/saveConfig
  config-command.ts (73) — buildCommand, buildCommandInDir, getEnvVars

All existing import paths unchanged (barrel re-exports preserve API surface).
Tests: 884 pass / 0 fail (unchanged from pre-split baseline).
