+++ src/commands/plugins/soul-sync/impl.ts
@@ -1,197 +1,22 @@
-import { existsSync, readdirSync, copyFileSync, mkdirSync, appendFileSync } from "fs";
+import { existsSync } from "fs";
 import { join, basename } from "path";
 import { hostExec } from "../../../sdk";
 import { loadConfig } from "../../../config";
-import { loadFleet, type FleetSession } from "../../shared/fleet-load";
+import { findPeers, findProjectsForOracle, syncOracleVaults, syncProjectVault, reportProjectResult, type SoulSyncResult, type ProjectSyncResult } from "./sync-helpers";
+import { resolveOraclePath, resolveProjectSlug, findOracleForProject } from "./resolve";
 
-const SYNC_DIRS = ["memory/learnings", "memory/retrospectives", "memory/traces"];
-
-/**
- * Sync new files from src dir to dst dir (skip existing).
- * Returns count of files copied.
- */
-export function syncDir(srcDir: string, dstDir: string): number {
-  if (!existsSync(srcDir)) return 0;
-  let count = 0;
-
-  function walk(src: string, dst: string) {
-    let entries: string[];
-    try { entries = readdirSync(src, { withFileTypes: true } as any) as any; }
-    catch { return; }
-
-    for (const entry of entries as any[]) {
-      const srcPath = join(src, entry.name);
-      const dstPath = join(dst, entry.name);
-      if (entry.isDirectory()) {
-        walk(srcPath, dstPath);
-      } else if (!existsSync(dstPath)) {
-        try {
-          mkdirSync(dst, { recursive: true });
-          copyFileSync(srcPath, dstPath);
-          count++;
-        } catch { /* skip unreadable files */ }
-      }
-    }
-  }
-
-  walk(srcDir, dstDir);
-  return count;
-}
+export { syncDir, findPeers, findProjectsForOracle, syncProjectVault } from "./sync-helpers";
+export type { SoulSyncResult, ProjectSyncResult } from "./sync-helpers";
+export { resolveOraclePath, resolveProjectSlug, findOracleForProject } from "./resolve";
 
 /**
- * Resolve ghq path for an oracle name.
- * Tries: ghq list --full-path | grep -i '/<stem>-oracle$'
- *
- * Defensive — accepts both bare oracle name ("neo") and full repo name
- * ("neo-oracle"). Strips trailing "-oracle" before re-appending so callers
- * passing either form land on the same lookup. (#372)
- */
-export async function resolveOraclePath(name: string): Promise<string | null> {
-  // Strip trailing -oracle so "neo" and "neo-oracle" both resolve identically.
-  const stem = name.replace(/-oracle$/, "");
-  try {
-    const out = await hostExec(`ghq list --full-path | grep -i '/${stem}-oracle$' | head -1`);
-    if (out?.trim()) return out.trim();
-  } catch { /* not found */ }
-
-  // Fallback: check fleet config for repo path
-  const ghqRoot = loadConfig().ghqRoot;
-  const fleet = loadFleet();
-  for (const sess of fleet) {
-    const oracleName = sess.name.replace(/^\d+-/, "");
-    if (oracleName === stem && sess.windows.length > 0) {
-      const repoPath = join(ghqRoot, sess.windows[0].repo);
-      if (existsSync(repoPath)) return repoPath;
-    }
-  }
-
-  return null;
-}
-
-/**
- * Find peer oracle names for a given oracle from fleet config.
- * Flat lookup — each oracle declares its own sync_peers.
- */
-export function findPeers(oracleName: string): string[] {
-  const fleet = loadFleet();
-  for (const sess of fleet) {
-    const name = sess.name.replace(/^\d+-/, "");
-    if (name === oracleName && sess.sync_peers) return sess.sync_peers;
-  }
-  return [];
-}
-
-/**
- * Find project repos this oracle absorbs from.
- */
-export function findProjectsForOracle(oracleName: string): string[] {
-  const fleet = loadFleet();
-  for (const sess of fleet) {
-    const name = sess.name.replace(/^\d+-/, "");
-    if (name === oracleName) return sess.project_repos || [];
-  }
-  return [];
-}
-
-/**
- * Resolve a git repo path to its canonical "org/repo" slug for `project_repos`
- * lookup. Handles both shapes of `ghqRoot`:
- *
- *   A. github.com-rooted: `/home/neo/Code/github.com`
- *      repoRoot = `/home/neo/Code/github.com/Soul-Brews-Studio/maw-js`
- *      → rel = `Soul-Brews-Studio/maw-js` → slug = `Soul-Brews-Studio/maw-js`
- *
- *   B. bare ghq root:     `/home/neo/Code`
- *      repoRoot = `/home/neo/Code/github.com/Soul-Brews-Studio/maw-js`
- *      → rel = `github.com/Soul-Brews-Studio/maw-js` → (strip host) →
- *        `Soul-Brews-Studio/maw-js` → slug = `Soul-Brews-Studio/maw-js`
- *
- * Before this normalization, shape B produced the org-only slug
- * `github.com/Soul-Brews-Studio` because `.split("/").slice(0, 2)` grabbed
- * the host + org instead of org + repo. See #193.
- *
- * Worktree suffix (`.wt-*`) is stripped from the repo segment so worktrees
- * match their parent repo's `project_repos` entry.
- *
- * Returns null if `repoRoot` is not under `ghqRoot` or doesn't have the
- * expected depth (e.g. sitting directly under ghqRoot with no org segment).
- */
-export function resolveProjectSlug(repoRoot: string, ghqRoot: string): string | null {
-  if (!repoRoot.startsWith(ghqRoot)) return null;
-  let rel = repoRoot.slice(ghqRoot.length).replace(/^\/+/, "");
-  // If ghqRoot is the bare ghq root (not github.com-rooted), rel starts with
-  // a host segment — strip known forges so we always land at "<org>/<repo>".
-  rel = rel.replace(/^(github\.com|gitlab\.com|bitbucket\.org)\//, "");
-  const parts = rel.split("/").slice(0, 2);
-  if (parts.length !== 2 || !parts[0] || !parts[1]) return null;
-  parts[1] = parts[1].replace(/\.wt-.*$/, "");
-  return parts.join("/");
-}
-
-/**
- * Find the oracle that owns a given project repo (org/repo slug).
- */
-export function findOracleForProject(projectRepo: string): string | null {
-  const fleet = loadFleet();
-  for (const sess of fleet) {
-    if (sess.project_repos?.includes(projectRepo)) {
-      return sess.name.replace(/^\d+-/, "");
-    }
-  }
-  return null;
-}
-
-export interface SoulSyncResult {
-  from: string;
-  to: string;
-  synced: Record<string, number>;
-  total: number;
-}
-
-/**
- * Sync ψ/memory/ from one oracle repo to another (new files only).
- */
-function syncOracleVaults(fromPath: string, toPath: string, fromName: string, toName: string): SoulSyncResult {
-  const fromVault = join(fromPath, "ψ");
-  const toVault = join(toPath, "ψ");
-
-  const synced: Record<string, number> = {};
-  for (const subdir of SYNC_DIRS) {
-    const src = join(fromVault, subdir);
-    const dst = join(toVault, subdir);
-    const count = syncDir(src, dst);
-    if (count > 0) synced[subdir] = count;
-  }
-
-  const total = Object.values(synced).reduce((a, b) => a + b, 0);
-
-  // Write sync receipt log
-  if (total > 0) {
-    const logDir = join(toVault, ".soul-sync");
-    try {
-      mkdirSync(logDir, { recursive: true });
-      const ts = new Date().toISOString();
-      const logLine = `${ts} | ${fromName} → ${toName} | ${total} files | ${Object.entries(synced).map(([k, v]) => `${v} ${k.split("/").pop()}`).join(", ")}\n`;
-      appendFileSync(join(logDir, "sync.log"), logLine);
-    } catch { /* non-critical */ }
-  }
-
-  return { from: fromName, to: toName, synced, total };
-}
-
-/**
- * maw soul-sync [peer] [--from <peer>]
- *
- * Flat mycelium model — any oracle syncs to any peer.
- *
- *   maw ss              push to all configured sync_peers
- *   maw ss <peer>       push to specific peer
- *   maw ss --from <p>   pull from specific peer
+ * maw soul-sync [peer] [--from <peer>] — push ψ/ to peers; --from reverses direction.
+ * maw ss <peer>       push to specific peer
+ * maw ss --from <p>   pull from specific peer
  */
 export async function cmdSoulSync(target?: string, opts?: { from?: boolean; cwd?: string }): Promise<SoulSyncResult[]> {
   const results: SoulSyncResult[] = [];
 
-  // Resolve current oracle
   let cwd = opts?.cwd || "";
   if (!cwd) {
     try {
@@ -205,7 +30,6 @@ export async function cmdSoulSync(target?: string, opts?: { from?: boolean; cwd?
   const repoName = cwdParts.pop() || "";
   const oracleName = repoName.replace(/-oracle$/, "").replace(/\.wt-.*$/, "");
 
-  // Resolve current oracle path (may be worktree, use git common dir)
   let oraclePath = cwd;
   try {
     const commonDir = (await hostExec(`git -C '${cwd}' rev-parse --git-common-dir`)).trim();
@@ -215,7 +39,6 @@ export async function cmdSoulSync(target?: string, opts?: { from?: boolean; cwd?
     }
   } catch { /* use cwd */ }
 
-  // Determine peers to sync with
   const peers = target ? [target] : findPeers(oracleName);
   if (peers.length === 0) {
     console.log(`  \x1b[33m⚠\x1b[0m soul-sync: no sync_peers configured for '${oracleName}'`);
@@ -261,70 +84,10 @@ export async function cmdSoulSync(target?: string, opts?: { from?: boolean; cwd?
   return results;
 }
 
-export interface ProjectSyncResult {
-  project: string;
-  oracle: string;
-  synced: Record<string, number>;
-  total: number;
-}
-
 /**
- * Sync ψ/memory/ from a project repo into an oracle repo.
- * Knowledge flows inward through the membrane — project → oracle, new files only.
- */
-export function syncProjectVault(
-  projectPath: string,
-  oraclePath: string,
-  projectRepo: string,
-  oracleName: string,
-): ProjectSyncResult {
-  const projectVault = join(projectPath, "ψ");
-  const oracleVault = join(oraclePath, "ψ");
-
-  const synced: Record<string, number> = {};
-  for (const subdir of SYNC_DIRS) {
-    const src = join(projectVault, subdir);
-    const dst = join(oracleVault, subdir);
-    const count = syncDir(src, dst);
-    if (count > 0) synced[subdir] = count;
-  }
-  const total = Object.values(synced).reduce((a, b) => a + b, 0);
-
-  if (total > 0) {
-    const logDir = join(oracleVault, ".soul-sync");
-    try {
-      mkdirSync(logDir, { recursive: true });
-      const ts = new Date().toISOString();
-      const logLine = `${ts} | project:${projectRepo} → ${oracleName} | ${total} files | ${Object.entries(synced).map(([k, v]) => `${v} ${k.split("/").pop()}`).join(", ")}\n`;
-      appendFileSync(join(logDir, "sync.log"), logLine);
-    } catch { /* non-critical */ }
-  }
-
-  return { project: projectRepo, oracle: oracleName, synced, total };
-}
-
-function reportProjectResult(r: ProjectSyncResult) {
-  if (r.total === 0) {
-    console.log(`  \x1b[90m○\x1b[0m project:${r.project} → ${r.oracle}: nothing new`);
-  } else {
-    const parts = Object.entries(r.synced).map(([dir, n]) => `${n} ${dir.split("/").pop()}`);
-    console.log(`  \x1b[32m✓\x1b[0m project:${r.project} → ${r.oracle}: ${parts.join(", ")}`);
-  }
-}
-
-/**
- * maw soul-sync --project
- *
- * Cell membrane absorbing nutrients — project ψ/ flows INWARD into oracle ψ/.
- *
- * Resolution (auto-detect by cwd):
- *   - Inside an oracle repo (name ends `-oracle`): walk fleet's `project_repos`,
- *     pull each project's ψ/ → my ψ/.
- *   - Inside a project repo: find the oracle whose `project_repos` lists this
- *     repo, push my ψ/ → that oracle's ψ/.
- *
- * Always direction = INWARD (project → oracle). Knowledge cannot flow outward;
- * project repos are the environment, the oracle is the cell.
+ * maw soul-sync --project — absorb project ψ/ into oracle ψ/ (inward only).
+ * Oracle cwd: absorbs from each project_repos entry.
+ * Project cwd: pushes ψ/ to the oracle that owns this repo.
  */
 export async function cmdSoulSyncProject(opts?: { cwd?: string }): Promise<ProjectSyncResult[]> {
   const results: ProjectSyncResult[] = [];
@@ -339,16 +102,13 @@ export async function cmdSoulSyncProject(opts?: { cwd?: string }): Promise<Proje
     }
   }
 
-  // Resolve git toplevel so we work even from a subdir
   let repoRoot = cwd;
   try {
     const top = (await hostExec(`git -C '${cwd}' rev-parse --show-toplevel`)).trim();
     if (top) repoRoot = top;
   } catch { /* not a git repo */ }
 
-  // Strip ghq root → "org/repo" slug. Drop worktree suffix for matching.
   const repoSlug = resolveProjectSlug(repoRoot, ghqRoot);
-
   const repoBase = basename(repoRoot).replace(/\.wt-.*$/, "");
   const isOracle = repoBase.endsWith("-oracle");
