+++ src/config.ts
@@ -1,542 +1,6 @@
-import { readFileSync, writeFileSync } from "fs";
-import { join } from "path";
-import { execSync } from "child_process";
-import { CONFIG_FILE } from "./core/paths";
-import { refreshContext } from "./lib/context";
-import { verbose, info } from "./cli/verbosity";
-
-function detectGhqRoot(): string {
-  try {
-    const root = execSync("ghq root", { encoding: "utf-8" }).trim();
-    // ghq may store repos under <root>/github.com/... — prefer that if it exists
-    const ghRoot = join(root, "github.com");
-    if (require("fs").existsSync(ghRoot)) return ghRoot;
-    return root;
-  } catch { return join(require("os").homedir(), "Code/github.com"); }
-}
-
-export type TriggerEvent = "issue-close" | "pr-merge" | "agent-idle" | "agent-wake" | "agent-crash" | "cron";
-
-export interface TriggerConfig {
-  on: TriggerEvent;
-  repo?: string;       // filter by repo (for issue-close, pr-merge)
-  timeout?: number;     // seconds (for agent-idle)
-  schedule?: string;    // crontab expression (for cron) — 5-field "m h dom mon dow"
-  action: string;       // shell command to execute — supports {agent}, {repo}, {issue} templates
-  name?: string;        // optional human label
-  once?: boolean;       // fire once then self-destruct (#149)
-}
-
-/** Named peer with URL */
-export interface PeerConfig {
-  name: string;
-  url: string;
-}
-
-export interface MawIntervals {
-  capture?: number;
-  sessions?: number;
-  status?: number;
-  teams?: number;
-  preview?: number;
-  peerFetch?: number;
-  crashCheck?: number;
-}
-
-export interface MawTimeouts {
-  http?: number;
-  health?: number;
-  ping?: number;
-  pty?: number;
-  workspace?: number;
-  shellInit?: number;
-  wakeRetry?: number;
-  wakeVerify?: number;
-}
-
-export interface MawLimits {
-  feedMax?: number;
-  feedDefault?: number;
-  feedHistory?: number;
-  logsMax?: number;
-  logsDefault?: number;
-  logsTruncate?: number;
-  messageTruncate?: number;
-  ptyCols?: number;
-  ptyRows?: number;
-}
-
-export interface MawConfig {
-  host: string;
-  port: number;
-  ghqRoot: string;
-  oracleUrl: string;
-  env: Record<string, string>;
-  commands: Record<string, string>;
-  sessions: Record<string, string>;
-  tmuxSocket?: string;
-  peers?: string[];
-  idleTimeoutMinutes?: number;
-  federationToken?: string;
-  autoRestart?: boolean;
-  triggers?: TriggerConfig[];
-  /** Node identity (e.g. "white", "mba") */
-  node?: string;
-  /** Named peers with URLs */
-  namedPeers?: PeerConfig[];
-  /** Agent → node mapping (e.g. { "homekeeper": "mba", "neo": "white" }) */
-  agents?: Record<string, string>;
-  /** GitHub org for maw bud (default: Soul-Brews-Studio) */
-  githubOrg?: string;
-  /** GitHub orgs to scan for oracle repos (default: Soul-Brews-Studio, laris-co) */
-  githubOrgs?: string[];
-  /** Fixed Claude session UUIDs per agent */
-  sessionIds?: Record<string, string>;
-  /** Path to ψ/ directory */
-  psiPath?: string;
-  /** TLS cert/key paths */
-  tls?: { cert: string; key: string };
-  /** Polling intervals (ms) */
-  intervals?: MawIntervals;
-  /** HTTP/operation timeouts (ms) */
-  timeouts?: MawTimeouts;
-  /** Buffer/display limits */
-  limits?: MawLimits;
-  /** HMAC auth window (seconds) */
-  hmacWindowSeconds?: number;
-  /** PIN for web UI */
-  pin?: string;
-  /** Plugin source URLs — auto-installed on bootstrap (nuke → first run) */
-  pluginSources?: string[];
-  /** Plugin names to disable (skip during scanning and execution) */
-  disabledPlugins?: string[];
-}
-
-const DEFAULTS: MawConfig = {
-  host: "local",
-  port: 3456,
-  ghqRoot: detectGhqRoot(),
-  oracleUrl: "http://localhost:47779",
-  env: {},
-  commands: { default: "claude" },
-  sessions: {},
-};
-
-/** Typed defaults for intervals, timeouts, limits (#172) */
-export const D = {
-  intervals: { capture: 50, sessions: 5000, status: 3000, teams: 3000, preview: 2000, peerFetch: 10000, crashCheck: 30000 } as const,
-  timeouts: { http: 5000, health: 3000, ping: 5000, pty: 5000, workspace: 5000, shellInit: 3000, wakeRetry: 500, wakeVerify: 3000 } as const,
-  limits: { feedMax: 500, feedDefault: 50, feedHistory: 50, logsMax: 500, logsDefault: 50, logsTruncate: 500, messageTruncate: 100, ptyCols: 500, ptyRows: 200 } as const,
-  hmacWindowSeconds: 300,
-} as const;
-
-/** Get a config interval with typed default fallback */
-export function cfgInterval(key: keyof typeof D.intervals): number {
-  return loadConfig().intervals?.[key] ?? D.intervals[key];
-}
-
-/** Get a config timeout with typed default fallback */
-export function cfgTimeout(key: keyof typeof D.timeouts): number {
-  return loadConfig().timeouts?.[key] ?? D.timeouts[key];
-}
-
-/** Get a config limit with typed default fallback */
-export function cfgLimit(key: keyof typeof D.limits): number {
-  return loadConfig().limits?.[key] ?? D.limits[key];
-}
-
-/** Get a top-level config value with default fallback */
-export function cfg<K extends keyof MawConfig>(key: K): MawConfig[K] {
-  return loadConfig()[key] ?? (DEFAULTS as MawConfig)[key];
-}
-
-let cached: MawConfig | null = null;
-
-/** Validate config values, warn on invalid fields, return sanitized config */
-function validateConfig(raw: Record<string, unknown>): Partial<MawConfig> {
-  const result: Record<string, unknown> = {};
-  const warn = (field: string, msg: string) =>
-    console.warn(`[maw] config warning: ${field} ${msg}, using default`);
-
-  // host: string, non-empty
-  if ("host" in raw) {
-    if (typeof raw.host === "string" && raw.host.trim().length > 0) {
-      result.host = raw.host.trim();
-    } else {
-      warn("host", "must be a non-empty string");
-    }
-  }
-
-  // port: number, 1-65535
-  if ("port" in raw) {
-    const p = Number(raw.port);
-    if (Number.isInteger(p) && p >= 1 && p <= 65535) {
-      result.port = p;
-    } else {
-      warn("port", "must be an integer 1-65535");
-    }
-  }
-
-  // ghqRoot: string
-  if ("ghqRoot" in raw) {
-    if (typeof raw.ghqRoot === "string" && raw.ghqRoot.length > 0) {
-      result.ghqRoot = raw.ghqRoot;
-    } else {
-      warn("ghqRoot", "must be a non-empty string");
-    }
-  }
-
-  // oracleUrl: string
-  if ("oracleUrl" in raw) {
-    if (typeof raw.oracleUrl === "string" && raw.oracleUrl.length > 0) {
-      result.oracleUrl = raw.oracleUrl;
-    } else {
-      warn("oracleUrl", "must be a non-empty string");
-    }
-  }
-
-  // env: Record<string, string>
-  if ("env" in raw) {
-    if (raw.env && typeof raw.env === "object" && !Array.isArray(raw.env)) {
-      result.env = raw.env;
-    } else {
-      warn("env", "must be an object");
-    }
-  }
-
-  // commands: Record<string, string>, must have "default" if present
-  if ("commands" in raw) {
-    if (raw.commands && typeof raw.commands === "object" && !Array.isArray(raw.commands)) {
-      const cmds = raw.commands as Record<string, unknown>;
-      if (!("default" in cmds) || typeof cmds.default !== "string") {
-        warn("commands", "must include a 'default' string entry");
-      } else {
-        result.commands = cmds as Record<string, string>;
-      }
-    } else {
-      warn("commands", "must be an object");
-    }
-  }
-
-  // sessions: Record<string, string>
-  if ("sessions" in raw) {
-    if (raw.sessions && typeof raw.sessions === "object" && !Array.isArray(raw.sessions)) {
-      result.sessions = raw.sessions;
-    } else {
-      warn("sessions", "must be an object");
-    }
-  }
-
-  // tmuxSocket: string if present
-  if ("tmuxSocket" in raw) {
-    if (typeof raw.tmuxSocket === "string") {
-      result.tmuxSocket = raw.tmuxSocket;
-    } else {
-      warn("tmuxSocket", "must be a string");
-    }
-  }
-
-  // triggers: TriggerConfig[]
-  if ("triggers" in raw) {
-    if (Array.isArray(raw.triggers)) {
-      const valid = raw.triggers.filter((t: any) => {
-        if (!t || typeof t !== "object") return false;
-        if (!t.on || typeof t.on !== "string") return false;
-        if (!t.action || typeof t.action !== "string") return false;
-        return true;
-      });
-      if (valid.length !== raw.triggers.length) {
-        warn("triggers", `has ${raw.triggers.length - valid.length} invalid entries, keeping valid ones`);
-      }
-      result.triggers = valid;
-    } else {
-      warn("triggers", "must be an array");
-    }
-  }
-
-  // federationToken: string, min 16 chars
-  if ("federationToken" in raw) {
-    if (typeof raw.federationToken === "string" && raw.federationToken.length >= 16) {
-      result.federationToken = raw.federationToken;
-    } else if (typeof raw.federationToken === "string") {
-      warn("federationToken", "must be at least 16 characters");
-    } else {
-      warn("federationToken", "must be a string");
-    }
-  }
-
-  // pin: string if present
-  if ("pin" in raw) {
-    if (typeof raw.pin === "string") {
-      result.pin = raw.pin;
-    } else {
-      warn("pin", "must be a string");
-    }
-  }
-
-  // pluginSources: string[] of URLs
-  if ("pluginSources" in raw) {
-    if (Array.isArray(raw.pluginSources)) {
-      result.pluginSources = (raw.pluginSources as unknown[]).filter(s => typeof s === "string") as string[];
-    } else {
-      warn("pluginSources", "must be an array of URL strings");
-    }
-  }
-
-  // disabledPlugins: string[] of plugin names
-  if ("disabledPlugins" in raw) {
-    if (Array.isArray(raw.disabledPlugins)) {
-      result.disabledPlugins = (raw.disabledPlugins as unknown[]).filter(s => typeof s === "string") as string[];
-    } else {
-      warn("disabledPlugins", "must be an array of plugin names");
-    }
-  }
-
-  // node: string if present
-  if ("node" in raw) {
-    if (typeof raw.node === "string" && raw.node.trim().length > 0) {
-      result.node = raw.node.trim();
-    } else {
-      warn("node", "must be a non-empty string");
-    }
-  }
-
-  // namedPeers: array of {name, url} objects
-  if ("namedPeers" in raw) {
-    if (Array.isArray(raw.namedPeers)) {
-      const valid = raw.namedPeers.filter((p: any) => {
-        if (!p || typeof p !== "object") return false;
-        if (typeof p.name !== "string" || typeof p.url !== "string") return false;
-        try { new URL(p.url); return true; } catch { return false; }
-      });
-      if (valid.length !== raw.namedPeers.length) {
-        warn("namedPeers", `has ${raw.namedPeers.length - valid.length} invalid entries`);
-      }
-      result.namedPeers = valid;
-    } else {
-      warn("namedPeers", "must be an array of {name, url}");
-    }
-  }
-
-  // agents: Record<string, string> (agent name → node name)
-  if ("agents" in raw) {
-    if (raw.agents && typeof raw.agents === "object" && !Array.isArray(raw.agents)) {
-      result.agents = raw.agents;
-    } else {
-      warn("agents", "must be an object mapping agent names to node names");
-    }
-  }
-
-  // peers: array of valid URLs if present
-  if ("peers" in raw) {
-    if (Array.isArray(raw.peers)) {
-      const valid = raw.peers.filter((p) => {
-        if (typeof p !== "string") return false;
-        try { new URL(p); return true; } catch { return false; }
-      });
-      if (valid.length !== raw.peers.length) {
-        warn("peers", `has ${raw.peers.length - valid.length} invalid URL(s), keeping valid ones`);
-      }
-      result.peers = valid;
-    } else {
-      warn("peers", "must be an array of URLs");
-    }
-  }
-
-  // githubOrg: string (#204)
-  if ("githubOrg" in raw && typeof raw.githubOrg === "string") {
-    result.githubOrg = raw.githubOrg;
-  }
-
-  // nanoclaw: pass through (bridge config)
-  if ("nanoclaw" in raw && raw.nanoclaw && typeof raw.nanoclaw === "object") {
-    result.nanoclaw = raw.nanoclaw;
-  }
-
-  return result as Partial<MawConfig>;
-}
-
-export function loadConfig(): MawConfig {
-  if (cached) return cached;
-  try {
-    const raw = JSON.parse(readFileSync(CONFIG_FILE, "utf-8"));
-    const validated = validateConfig(raw);
-    cached = { ...DEFAULTS, ...validated };
-  } catch {
-    cached = { ...DEFAULTS };
-  }
-  // One-shot startup summary — fires unless --quiet/--silent (verbose-by-default).
-  verbose(() => {
-    const nT = cached!.triggers?.length ?? 0;
-    const nP = cached!.pluginSources?.length ?? 0;
-    const nPeers = (cached!.peers?.length ?? 0) + (cached!.namedPeers?.length ?? 0);
-    info(`loaded config: ${nT} trigger${nT === 1 ? "" : "s"}, ${nP} plugin${nP === 1 ? "" : "s"}, ${nPeers} peer${nPeers === 1 ? "" : "s"}`);
-  });
-  return cached;
-}
-
-/** Reset cached config (for hot-reload or testing) */
-export function resetConfig() {
-  cached = null;
-}
-
-/** Write config to maw.config.json and reset cache */
-export function saveConfig(update: Partial<MawConfig>) {
-  const current = loadConfig();
-  const merged = { ...current, ...update };
-  writeFileSync(CONFIG_FILE, JSON.stringify(merged, null, 2) + "\n", "utf-8");
-  resetConfig(); // clear cache so next loadConfig() reads fresh
-  refreshContext(); // clear DI cache so middleware picks up new config
-  return loadConfig();
-}
-
-/** Validate config shape with native TS checks (no Zod).
- *  Returns array of error strings — empty means valid. */
-export function validateConfigShape(config: unknown): string[] {
-  const errors: string[] = [];
-  if (!config || typeof config !== "object") return ["Config must be an object"];
-  const c = config as Record<string, unknown>;
-
-  if (c.host !== undefined && typeof c.host !== "string") errors.push("host must be a string");
-  if (c.port !== undefined) {
-    if (typeof c.port !== "number" || !Number.isInteger(c.port) || c.port < 1 || c.port > 65535)
-      errors.push("port must be an integer 1-65535");
-  }
-  if (c.ghqRoot !== undefined && typeof c.ghqRoot !== "string") errors.push("ghqRoot must be a string");
-  if (c.oracleUrl !== undefined && typeof c.oracleUrl !== "string") errors.push("oracleUrl must be a string");
-  if (c.tmuxSocket !== undefined && typeof c.tmuxSocket !== "string") errors.push("tmuxSocket must be a string");
-  if (c.federationToken !== undefined && typeof c.federationToken !== "string") errors.push("federationToken must be a string");
-
-  if (c.env !== undefined) {
-    if (!c.env || typeof c.env !== "object" || Array.isArray(c.env)) {
-      errors.push("env must be a Record<string, string>");
-    } else {
-      for (const [k, v] of Object.entries(c.env as Record<string, unknown>)) {
-        if (typeof v !== "string") errors.push(`env.${k} must be a string`);
-      }
-    }
-  }
-
-  if (c.commands !== undefined) {
-    if (!c.commands || typeof c.commands !== "object" || Array.isArray(c.commands)) {
-      errors.push("commands must be a Record<string, string>");
-    } else {
-      for (const [k, v] of Object.entries(c.commands as Record<string, unknown>)) {
-        if (typeof v !== "string") errors.push(`commands.${k} must be a string`);
-      }
-    }
-  }
-
-  if (c.sessions !== undefined) {
-    if (!c.sessions || typeof c.sessions !== "object" || Array.isArray(c.sessions)) {
-      errors.push("sessions must be a Record<string, string>");
-    } else {
-      for (const [k, v] of Object.entries(c.sessions as Record<string, unknown>)) {
-        if (typeof v !== "string") errors.push(`sessions.${k} must be a string`);
-      }
-    }
-  }
-
-  if (c.peers !== undefined) {
-    if (!Array.isArray(c.peers)) {
-      errors.push("peers must be a string[]");
-    } else {
-      for (let i = 0; i < c.peers.length; i++) {
-        if (typeof c.peers[i] !== "string") errors.push(`peers[${i}] must be a string`);
-      }
-    }
-  }
-
-  return errors;
-}
-
-/** Return config with env values masked for display */
-export function configForDisplay(): MawConfig & { envMasked: Record<string, string> } {
-  const config = loadConfig();
-  const envMasked: Record<string, string> = {};
-  for (const [k, v] of Object.entries(config.env)) {
-    if (v.length <= 4) {
-      envMasked[k] = "\u2022".repeat(v.length);
-    } else {
-      envMasked[k] = v.slice(0, 3) + "\u2022".repeat(Math.min(v.length - 3, 20));
-    }
-  }
-  const result: any = { ...config, env: {}, envMasked };
-  // Mask federation token (show first 4 chars only)
-  if (result.federationToken) {
-    result.federationToken = result.federationToken.slice(0, 4) + "\u2022".repeat(12);
-  }
-  return result;
-}
-
-/** Simple glob match: supports * at start/end (e.g., "*-oracle", "codex-*") */
-function matchGlob(pattern: string, name: string): boolean {
-  if (pattern === name) return true;
-  if (pattern.startsWith("*") && name.endsWith(pattern.slice(1))) return true;
-  if (pattern.endsWith("*") && name.startsWith(pattern.slice(0, -1))) return true;
-  return false;
-}
-
-/** Build the full command string for an agent (no env vars — use setSessionEnv) */
-export function buildCommand(agentName: string): string {
-  const config = loadConfig();
-  let cmd = config.commands.default || "claude";
-
-  // Strip --dangerously-skip-permissions when running as root (#181)
-  if (process.getuid?.() === 0) {
-    cmd = cmd.replace(/\s*--dangerously-skip-permissions\b/, "");
-  }
-
-  // Match specific patterns first (skip "default")
-  for (const [pattern, command] of Object.entries(config.commands)) {
-    if (pattern === "default") continue;
-    if (matchGlob(pattern, agentName)) { cmd = command; break; }
-  }
-
-  // Inject --session-id if configured for this agent
-  const sessionIds: Record<string, string> = (config as any).sessionIds || {};
-  const sessionId = sessionIds[agentName]
-    || Object.entries(sessionIds).find(([p]) => p !== "default" && matchGlob(p, agentName))?.[1];
-  if (sessionId) {
-    // Use --resume with fixed session ID (--session-id locks, --resume doesn't)
-    // Replace --continue with --resume <uuid> if present, otherwise append
-    if (cmd.includes("--continue")) {
-      cmd = cmd.replace(/\s*--continue\b/, ` --resume "${sessionId}"`);
-    } else {
-      cmd += ` --resume "${sessionId}"`;
-    }
-  }
-
-  // Prefix: load direnv + clear stale CLAUDECODE.
-  // direnv allow + export ensures .envrc env vars load before Claude starts,
-  // since tmux send-keys can race with the shell's direnv hook.
-  // If direnv is not installed, `direnv allow` fails visibly (diagnostic),
-  // && short-circuits, and the rest of the block runs normally.
-  // unset CLAUDECODE prevents "cannot be launched inside another" from crashed sessions.
-  const prefix = "direnv allow . && eval \"$(direnv export zsh)\"; unset CLAUDECODE;";
-
-  // If command uses --continue or --resume, add shell fallback without it.
-  // --continue errors when no prior conversation exists (e.g. fresh worktree,
-  // wiped session). --resume errors when session ID doesn't exist yet.
-  // The fallback retries the same command minus --continue/--resume,
-  // but keeps --session-id if present so the first run creates the session with that ID.
-  if (cmd.includes("--continue") || cmd.includes("--resume")) {
-    let fallback = cmd.replace(/\s*--continue\b/, "").replace(/\s*--resume\s+"[^"]*"/, "");
-    if (sessionId) fallback += ` --session-id "${sessionId}"`;
-    return `${prefix} ${cmd} || ${prefix} ${fallback}`;
-  }
-
-  return `${prefix} ${cmd}`;
-}
-
-/** Wrap buildCommand with cd to ensure correct working directory after reboot.
- *  Parenthesize buildCommand so cd applies to both primary + fallback in `cmd || fallback`.
- *  Otherwise shell precedence (`&&` tighter than `||`) makes the fallback run without cd. */
-export function buildCommandInDir(agentName: string, cwd: string): string {
-  return `cd '${cwd}' && { ${buildCommand(agentName)}; }`;
-}
-
-/** Get env vars from config (for tmux set-environment) */
-export function getEnvVars(): Record<string, string> {
-  return loadConfig().env || {};
-}
+// Barrel — re-exports everything so all existing import paths remain unchanged.
+export type { TriggerEvent, TriggerConfig, PeerConfig, MawIntervals, MawTimeouts, MawLimits, MawConfig } from "./config-types";
+export { D } from "./config-types";
+export { validateConfigShape } from "./config-validate";
+export { loadConfig, resetConfig, saveConfig, configForDisplay, cfgInterval, cfgTimeout, cfgLimit, cfg } from "./config-load";
+export { buildCommand, buildCommandInDir, getEnvVars } from "./config-command";
